backend/auth/jwt_handler.py
from jose import jwt
from datetime import datetime, timedelta
from backend.config import JWT_SECRET, JWT_ALGORITHM, TOKEN_EXPIRE_MINUTES

def create_token(user_id: int):
    payload = {
        "user_id": user_id,
        "exp": datetime.utcnow() + timedelta(minutes=TOKEN_EXPIRE_MINUTES)
    }
    return jwt.encode(payload, JWT_SECRET, algorithm=JWT_ALGORITHM)

def decode_token(token: str):
    return jwt.decode(token, JWT_SECRET, algorithms=[JWT_ALGORITHM])
