backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
passlib[bcrypt]
argon2_cffi
requests
pandas